[fix] move all java code from src/main/scala to src/main/java
